requirements.txt
line-bot-sdk
flask
googletrans
pyjokes